refactor: move parsing to own class
